refactor: replace MatchQueryBuilder with direct matchesResource calls in EventQueryBuilder and remove match-builder files
