+++ packages/tba-sdk/src/builders/index.ts
@@ -1,5 +1,4 @@
 export { EventQueryBuilder } from "./event-builder";
-export { MatchQueryBuilder } from "./match-builder";
 export { TeamQueryBuilder } from "./team-builder";
 
 export type { CompLevel, MatchQueryFilters } from "./types";
